+++ src/main/java/com/cjw/reactivecommunityproject/common/exception/model/RcBaseException.java
@@ -4,7 +4,7 @@ package com.cjw.reactivecommunityproject.common.exception.model;
 import org.springframework.http.HttpStatus;
 
 public abstract class RcBaseException extends RuntimeException implements RcBaseErrorMessage {
-    protected Long errorCode;
+    protected Integer errorCode;
     protected String errorMessage;
     protected HttpStatus httpStatus;
     protected Boolean isDisplay;
